libGL: use headers from glvnd
The glvnd library contains a copy of the OpenGL headers.  By compiling
against the glvnd headers we completely avoid a dependency on Mesa (and
LLVM).  This reduces rebuild time.  It also prevents accidents, such as
where some packages kept a reference to the mesa.drivers output and thus
had all Mesa drivers in their runtime closure.