pythonPackages.hetzner: 0.8.2 -> 0.8.3
Recent changes in the Hetzner Robot API have removed a few obsolete
fields which version 0.8.2 was still referencing and which is now fixed
in version 0.8.3.

Due to a misunderstanding on my side I haven't updated to version 0.8.3
in nixpkgs yet, which resulted in this delay.

This fixes the NixOps Hetzner backend.

Signed-off-by: aszlig <aszlig@nix.build>