pythonPackages.greenlet: 0.4.17 -> 1.0.0
Update greenlet to 1.0.0.

meinheld declares a dependency to greenlet <0.5, but according to [1]
and [2], there are no API or ABI changes leading to greenlet-1.0.0, so
it seems reasonable to drop the strict requirement.  This commit
uses substituteInPlace to patch setup.py since no upstream commit has
landed in the main branch yet.

Morehover, quoting [2]:

    Prior to greenlet 1.0 there were no semantic meanings attached to
    greenlet versions — API and ABI regularly changed from 0.4.x
    to 0.4.x+1, so their current pin doesn't make much sense anyway.

nix-review reveals few broken packages that are also broken in master.

[1] https://github.com/mopemope/meinheld/pull/123
[2] https://github.com/benoitc/gunicorn/issues/2541#issuecomment-800353993