[unpackfsc] Simplify to **one** unpack action
The whole getup with multiple unpack actions makes the module
more complicated; instead, use one unpack action only. If there
are more of them, use multiple instances.

Similarly, remove the Config object: The **view** modules need a
Config object, because the config can be used by multiple UIs, and
we don't want to mix UI with business logic. For a single job,
that's different. Just put the job config in the job itself.