cmake: Fix Darwin SDK module.
The preConfigure step applies a very broad replacement of /usr and /opt
to /var/empty. However, the replacement affects also paths where /usr
and /opt appear in the middle.

One of those cases was breaking the Darwin SDK CMake module. In that
module many of the paths are relative to the SDK root (which is hold in
CMake variables like `_CMAKE_OSX_SYSROOT_PATH`), which actually include
an /usr path, with headers and libraries inside. The replacement was
breaking those paths, and making very difficult to use CMake to target
macOS, or other Darwin platforms.

The changes modify the replacement to skip lines where the replacement
is preceeded by the variables used in the CMake files.