darwin.apple_sdk.frameworks: add missing frameworks required by pyobjc
pyobjc is effectively a bunch of Python bindings to a whole heap of the
frameworks in the Apple SDK - as such, we need to be able to declare
dependencies on a consistent set of frameworks as exposed by a particular macOS
SDK version.

Since the current SDK version fetched here is 10.12, this commit exports all
the frameworks that pyobjc expects for macOS 10.12.