python3Packages.pyobjc: 7.0.1 -> 7.1
At the same time, this actually fixes the build so it works at all; previously
it was both disabled on Python 3, and not working on Python 2.7.

pyobjc actually consists of a bunch of sub-modules (one for each wrapped
framework, give-or-take). For parallelising the build at maximum efficiency, as
well as allowing subpackages to select only the pyobjc dependencies which they
actually need, we package each of the submodules individually, rather than
pyobjc as a single coherent unit.

To make this easier, we use the GitHub src archive, and patch it once upfront,
which means we only need to move a single src pointer as we update (rather than
fetching from PyPI, which would entail having separate fetch directives for
each submodule).

Additional packaging notes:
* A lot of the exposed APIs are SDK-version-dependent, and nixpkgs takes great
  pains to expose separate frameworks from the SDK separately. In order for
  this to work, we patch out all the SDK version checks to instead use an
  environment variable with the macOS version, which we fix at 10.12 (which is,
  as of writing, the SDK version used).
* LaunchServices just appears to be broken, and this breaks importing the
  entire CoreServices module. That's patched too.
* iTunesLibrary is not actually a core macOS framework, but is usually provided
  in pyobjc. I don't want to delve into figuring out how to expose it properly
  as a framework, so for the purposes of getting a basic package up for review
  I've just dropped it.
* Given the large number of frameworks that are needed it's impractical to
  separately provide these as arguments to pyobjc, so the entire frameworks
  attrset is provided as an argument instead.