pkgs-lib/formats: support top level INI keys
lib/generators: support top level keys in toINI

Many packages using INI as their configuration format allow top-level
keys that don't belong to a section before the first section, for
example gmnisrv and foot. However lib.generators.toINI and
pkgs.formats.ini don't support this currently. This commit introduces
support for this.

All keys in the attribute set passed are considered as top-level entries
if they are not attribute sets (that is don't represent a section).
Introducing a switch for turning this behavior on or off doesn't seem
necessary as there is no valid representation for such values in the
input attribute set for the use case without top-level attributes.

Now such things are possible:

```nix
lib.generators.toINI {} {
  foo = 13;
  bar = {
    baz = 12;
    hello = "world";
  };
} == ''
  foo=13
  [bar]
  baz=12
  hello=world
''
```

Thanks to Ben Sima for cleaning up and porting a snippet I had for
proper nixpkgs!

Co-authored-by: Ben Sima <ben@bsima.me>