nixos/nix-daemon: assert system or systems for buildMachines.
Commit 5395397f removed the assertions from the buildMachines to
ensure that either system or systems is set for each buildmachine.

This patch re-implements those assertions.

The symptom is that if both system and systems are omitted, then the
/etc/machines file has the wrong number of columns and any attempt to
run a `nix` operation that has to perform a build will fail with a
`strtoull` exception.

(cherry picked from commit 58921a49040285b76da8575fc3dd3d41fc42f895)