pythonPackages.cxxfilt: init at 0.2.2
Adds the cxxfilt Python package for demangling C++ symbols in
Python (interfacing to abi::__cxa_demangle).

It appears to work with both Python 2 and Python 3, although I could
only test it using Python 3.

The package uses its own mechanism on top of ctypes for discovering a
C++ standard library. With Nix this does not seem to work though, and
hence the absolute path of the shared object is specified
instead. This should also make Nix pick up this library as a runtime
dependency of the package. A simple substituteInPlace of the
respective Python method invocation seems fine here, as using a proper
patch would've required a substitution for the path anyways.

Signed-off-by: Leon Schuermann <leon@is.currently.online>
Reviewed-by: William Casarin <jb55@jb55.com>
Tested-by: William Casarin <jb55@jb55.com>
Reviewed-by: Matthias Beyer <mail@beyermatthias.de>
Link: https://lists.sr.ht/~andir/nixpkgs-dev/%3C20210410113406.349-1-leon@is.currently.online%3E