hdf5_1_10: rollback 1.10.7 -> 1.10.6
This package only exists for compatibility with
``python3Packages.tables``, as far as I can tell based on usage in
nixpkgs.

``tables`` tests fail on hdf5 > v1.10.5, so we are reverting this to
the last "good" version of hdf5 that works with ``tables``.
``hdf5_1_10`` should hopefully be unpinned on the next release of
``python3Packages.tables``.
https://github.com/PyTables/PyTables/issues/845