ubootOdroidC2: Review build to produce a single image file
This is somwhat closer to what other U-Boot builds produce.

The end-user will need to flash the first 442 bytes in the MBR, but
that's what "sd_fusing.sh" did... mostly...

This removes the "sd_fusing.sh" script, but it wasn't good anyway, as it
only worked when CWD was the directory it was in, and forced sudo usage.

This was verified to produce the same exact binaries in these
conditions:

Before this change:

```
 $ cd result/
 $ ./sd_fusing.sh .../test.img
```

After this change:

```
dd if=result/bl1.bin.hardkernel of=test.img conv=fsync,notrunc bs=1 count=442
dd if=result/u-boot.img         of=test.img conv=fsync,notrunc bs=512 seek=1
```

So it is assumed this change is 1:1 compatible in the end.