nixos/tests/hibernate: use static netcat
This attempts to make this test a little bit more robust: it's
inherently a bit janky, because the Nix store is only available over 9P
and the mount doesn't get resumed properly across reboots, but this
tries to make the best of it.

I _think_ what was happening before was that under some situations some
parts of the netcat binary were being paged out, and the attempt to
demand-page them back in was leading to... problems. Sometimes the test
would pass, sometimes it wouldn't.

Switching to using a static binary, which is then explicitly copied to
and run from /tmp seems to make this test pass much more robustly - at
least, I haven't been able to make it fail after running it 100 times in
this configuration.