glibc: always compile with a gcc stdenv
Our glibc doesn't compile with clang and useLLVM doesn't make the stdenv
implicitly switch to musl (and arguably shouldn't do that). To make LLVM
cross compilation without patching glibc to compile with clang (which is
probably a lot of work), always compile glibc with gcc. This should
unbreak a lot of cross compilation tests, I have tested:

* tests.cross.llvm.gnu64
* tests.cross.llvm.aarch64-multiplatform

Basing against staging because this change only really makes sense if
useLLVM works with llvmPackages_7 which is only the case in staging at
the moment.