zfsUnstable: 2.0.4 -> 2.1.0-rc3
In https://github.com/openzfs/zfs/commit/658fb802 OpenZFS introduced a
new compatibility feature. This is accompanied by some default profiles
which are located in the $out output of the Nix derivation. However,
these compatibility profiles are used by both the zpool executable
located in $out and the libzfs library located in $lib. This
inadvertently creates a cycle, because zpool refers to $lib to link
against libzfs but libzfs refers back to $out for the compatibility
profiles.

There are two possible routes to rectify this problem:

 1. Patch the zpool source code to not read the compatibility profiles
    and adjust the build system to put them into the $lib output.

 2. Simply remove the $out/$lib split.

Since no other derivation in nixpkgs refers to the $lib output
explicitly we opt for the latter, because this is also in accordance
with upstream.