matrix-synapse: fix build for 20.09
I had to drop the strict requirement on cryptography here which is only
there to make sure that users have cryptography with a recent openssl.
This is not necessary in our case since we build version 3.2 with the
latest openssl available (also on 20.09).

A full description can be found in the commit message of the patchfile.