systemd: fix lxd support with ubuntu patches
Currently things are like this:

systemd: if /sys is writeable we can use namespaces
ubuntu: no, not on lxd.
so systemd fails to start things properly based on wrong assumptions.

currently there's no upstream patch for it, so ubuntu has one that makes it work. it makes sense to add it here so we're compatible with lxd.

So that's the patch and some others to fix startup errors in lxd