nixos/borgbackup: optionally run as DynamicUser
When setting user to `null` the systemd unit uses a DynamicUser setup
with the CAP_DAC_READ_SEARCH capability, thereby being able to bypass
filesystem read and search permission checks.

To still be able to achieve privileged pre and post tasks, they are
moved into ExecStartPre and ExecStopPost respectively.

This doesn't come without breaking changes. The custom `exitStatus`
variable is being dropped in favor of the inherited `EXIT_STATUS` and
`EXIT_CODE`, that are provided by systemd itself. The exit trap itself
seems to have only been created for cleanup tasks, which are moved to
ExecStopPost, and the `exitStatus` variable, so it has been dropped
entirely.