nixos/tests: update to work with DynamicUser change
The local job wants to write to /root/back:up, so it is crucial that it
runs as root, make that explicit.

The `archiveName` option cannot be modified in `preHook` any longer,
since it is not inserted into the backup script any longer. Instead we
parse the archive name dynamically from `borg list` now.

Converted the local jobset to f'-strings, to make them easier to parse.