python2 and python3: build unoptimized bytecode again, build no longer reproducible
In 9d03ff52229f36ff4b4e9d0ee2db4a46b90512bb I made the CPython builds
reproducible. This required not generating default unoptimized bytecode.
I was under the impression the optimized bytecode would be used then,
but you need to opt-in on that. Not having the default bytecode resulted
in a significant performance hit. Therefore, bytecode is generated again
in this commit, and thereby the builds are no longer reproducible.