nixos/systemd: Add overrides to unit derivations
This one might be a bit of a corner case but it really makes your life
easier when you need something like this. Basically, it allows adding
overrides to the derivation of unit files.

We have two use cases, both are related to out-of tree work. The first
is an out-of-tree AppArmor module that attaches profiles to the literal
unit files and loads them on activation. The second one is an override
to gitlab-runner which is needed since we run it in a really
non-standard way. It essentially removes the `+` from the ExecStartPre
line. This would not have been as easy without this module we already
had laying around. Maybe it will be of use for other people and use
cases as well.