cxxtest: build using python3
It is 2021 we should start moving everything to python3 :-)

I randomly saw this package on hydra and thought it might be a good candidate to port.

The project follows a very unique approach to supporting python3. The
python3 code is in a subdirectory of the python directory and the
setup.py changes into that directory during execution when python3 is
detected. This required changing a few directories in our build script
to match that layout.

Also the python package `ply` is now a dependency as they are using it
for parsing C++ code.

When reviewing this do not be confused by the first test result line
that says that zero tests have been ran. Tests are being executed in the
postCheck phase (after compiling the test binary).