vscode-extensions.ms-python.python: refactor
In order for the ms-python.python extension to work these days,
we also need the ms-toolsai.jupyter extension.  To permit easy
upgrades and to honour the principle of least surprise, we
wrapper the Python extension to include the Jupyter one.  The
reverse is not true, you *can* have the Jupyter extension without
the Python one, so the reverse is not also done.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>